pull up config validation graph
